refactor: move the shared-evaluations git adapter to data layer
data/fs/shared_repo.py owns the clone-mirror porcelain; the API keeps its
services/shared_repo facade (re-export subset), and service-layer callers
import the data module directly.
