+++ src/quodeq/services/shared_repo.py
@@ -1,368 +1,23 @@
-"""Shared results repository: clone, refresh, and path management.
+"""Shared-evaluations clone access for delivery layers.
 
-The shared repo is a git remote holding an evaluations/ tree in the same
-layout as the local evaluations dir. We keep a full clone (no --depth,
-results repos are small; a shallow clone made git-log-based attribution
-misattribute every project to whoever pushed last -- audit finding C1)
-under ~/.quodeq/cache/shared/<url-hash>/repo (QUODEQ_CACHE_ROOT overrides
-the base).
-
-Git-mutation serialization (audit finding C2): background refreshes and an
-in-flight publish share one clone directory, so an unserialized fetch +
-hard-reset racing a stage/commit/push can tear a commit or contend on
-.git/index.lock. `clone_lock()` gives every mutator (refresh_shared_clone,
-ensure_shared_clone, publish_project) one process-wide lock per clone path.
-Readers (published_meta, sync_shared_index, the read-only /api/shared/*
-route mirrors) are deliberately NOT serialized against this lock -- full
-reader serialization is deferred. Torn reads are bounded instead by Task 2's
-atomic meta writes (published.json via tmp+rename): a reader can see a
-project that is one publish stale, never a half-written published.json.
+The git-clone-mirror adapter itself lives in ``data/fs/shared_repo.py``; the
+API layer does not import ``data/`` directly (see ARCHITECTURE.md import
+rules), so the symbols its routes need are re-exported here. Service-layer
+code imports ``quodeq.data.fs.shared_repo`` directly.
 """
-from __future__ import annotations
-
-import hashlib
-import json
-import logging
-import os
-import shutil
-import stat
-import subprocess
-import threading
-from pathlib import Path
-
-# Re-exported so the api layer can validate a shared-repo URL without
-# importing quodeq.data directly (the import-layer gate in
-# tools/check_imports.py allows api -> services and services -> data, but
-# not api -> data). services/evaluation_mixin.py imports the same function
-# straight from quodeq.data.fs.repo_validation for the same reason.
-from quodeq.data.fs.repo_validation import validate_remote_url  # noqa: F401
-
-logger = logging.getLogger(__name__)
-
-_CACHE_ENV = "QUODEQ_CACHE_ROOT"
-_DEFAULT_GIT_TIMEOUT_S = 300
-
-
-def _clear_readonly_and_retry(func, path, exc):  # noqa: ARG001
-    """rmtree onexc callback for git trees. Git marks object files read-only,
-    and on Windows deleting a read-only file raises PermissionError (POSIX
-    deletion only checks the parent dir). Clear the bit and retry once; if
-    that also fails, log instead of raising so cleanup stays best-effort."""
-    try:
-        os.chmod(path, stat.S_IWRITE)
-        func(path)
-    except OSError as retry_exc:
-        logger.warning("failed to remove %s: %s", path, retry_exc)
-
-
-def remove_clone_dir(path: Path | str) -> None:
-    """Best-effort recursive delete that tolerates git's read-only files.
-
-    Missing paths are a no-op. Never raises: a half-removed or
-    permission-denied dir must not turn cleanup into a 500.
-    """
-    if not os.path.lexists(path):
-        return
-    shutil.rmtree(path, onexc=_clear_readonly_and_retry)
-
-
-def _git_env() -> dict[str, str]:
-    """Environment for git subprocess calls.
-
-    GIT_LFS_SKIP_SMUDGE avoids pulling LFS blobs we don't need. GIT_TERMINAL_PROMPT=0
-    stops git from blocking on an interactive credential or passphrase prompt, since
-    these subprocess calls have stdin closed (see run_git) and nobody is there to answer.
-
-    Known limitation: GIT_TERMINAL_PROMPT only covers prompts issued by git
-    itself. ssh reads from /dev/tty directly, so a first-contact host-key
-    confirmation or a key passphrase without a loaded agent still blocks, and
-    the call only dies at the run_git timeout. ssh remotes need the host in
-    known_hosts and the key in an agent (or use an https remote instead).
-    """
-    return {**os.environ, "GIT_LFS_SKIP_SMUDGE": "1", "GIT_TERMINAL_PROMPT": "0"}
-
-
-def run_git(
-    args: list[str], *, cwd: Path | None = None, timeout: int = _DEFAULT_GIT_TIMEOUT_S
-) -> tuple[bool, str]:
-    try:
-        proc = subprocess.run(
-            ["git", *args],
-            cwd=str(cwd) if cwd else None,
-            env=_git_env(),
-            stdin=subprocess.DEVNULL,
-            capture_output=True,
-            text=True,
-            encoding="utf-8",
-            errors="replace",
-            timeout=timeout,
-        )
-        return proc.returncode == 0, (proc.stdout or "") + (proc.stderr or "")
-    except (OSError, subprocess.TimeoutExpired) as exc:
-        return False, str(exc)
-
-
-def _cache_base(env: dict | None = None) -> Path:
-    e = env if env is not None else os.environ
-    base = e.get(_CACHE_ENV)
-    root = Path(base) if base else Path.home() / ".quodeq" / "cache"
-    return root / "shared"
-
-
-def shared_cache_dir(url: str, env: dict | None = None) -> Path:
-    digest = hashlib.sha256(url.strip().encode("utf-8")).hexdigest()[:16]
-    return _cache_base(env) / digest
-
-
-def shared_repo_path(url: str, env: dict | None = None) -> Path:
-    return shared_cache_dir(url, env) / "repo"
-
-
-def shared_evaluations_root(url: str, env: dict | None = None) -> Path:
-    return shared_repo_path(url, env) / "evaluations"
-
-
-_CLONE_LOCKS: dict[str, threading.RLock] = {}
-_CLONE_LOCKS_GUARD = threading.Lock()
-
-
-def clone_lock(url: str, env: dict | None = None) -> threading.RLock:
-    """Process-wide reentrant lock serializing git mutations on one clone.
-
-    Keyed by the clone's resolved path (shared_repo_path), so any two
-    callers that resolve to the same cache directory share one lock
-    regardless of how they spelled the url. RLock, not Lock: publish_project
-    holds this lock across its own internal refresh_shared_clone call, and
-    refresh_shared_clone / ensure_shared_clone each acquire it again inside
-    their own bodies -- a plain Lock would deadlock a thread against itself.
-    See the module docstring for what this does and does not serialize.
-    """
-    key = str(shared_repo_path(url, env))
-    with _CLONE_LOCKS_GUARD:
-        return _CLONE_LOCKS.setdefault(key, threading.RLock())
-
-
-def ensure_shared_clone(url: str, env: dict | None = None) -> Path | None:
-    with clone_lock(url, env):
-        repo = shared_repo_path(url, env)
-        if (repo / ".git").exists():
-            return repo
-        repo.parent.mkdir(parents=True, exist_ok=True)
-        ok, out = run_git(["clone", "--", url, str(repo)])
-        if not ok:
-            logger.warning("shared clone failed for %s: %s", url, out.strip()[:500])
-            remove_clone_dir(repo)
-            return None
-        return repo
-
-
-_DEFAULT_REFRESH_TIMEOUT_S = 30
-
-
-def refresh_shared_clone(
-    url: str, env: dict | None = None, *, timeout: int = _DEFAULT_REFRESH_TIMEOUT_S
-) -> tuple[bool, str]:
-    """Fetch + hard-reset the clone to the remote's HEAD.
-
-    Returns ``(ok, reason)``: *reason* is ``""`` on success, else the
-    failing git command's stderr/stdout tail (<=200 chars) -- callers (the
-    refresh route, the ?refresh=1 listing branch) surface this so a failed
-    refresh reads as "could not resolve host" or "authentication failed"
-    instead of a bare "Request failed: 502" that can't distinguish DNS vs
-    auth vs a deleted origin (audit finding B3). Every failure is ALSO
-    logged via logger.warning, so a background/best-effort caller that
-    discards *reason* (e.g. publish_project's internal refresh) still gets
-    a diagnosable server-side trail.
-
-    Called in-request (GET /api/shared/projects?refresh=1, POST
-    /api/shared/refresh), so it must not inherit run_git's 300s default --
-    a black-holed connection would otherwise hang the request for up to 5
-    minutes. *timeout* bounds both the fetch (the network call) and the
-    reset (local but kept consistent); a black-holed connection now turns
-    into a stale=true response in ~*timeout* seconds instead. Does not
-    affect ensure_shared_clone's own (still 300s) clone timeout -- an
-    initial clone can legitimately take much longer than a refresh.
-
-    Unshallowing only applies to NEW clones: ensure_shared_clone stopped
-    passing --depth 1 in a prior fix, but a shared-clone cache directory
-    created back when it still did stays shallow forever otherwise --
-    ensure_shared_clone early-returns because `.git` already exists, and a
-    plain `fetch origin HEAD` does not unshallow a repo on its own. If
-    `.git/shallow` is present, try `git fetch --unshallow origin` first; a
-    failure there (network hiccup, odd remote) is not fatal -- fall through
-    to the plain fetch below, and a later refresh call retries the unshallow.
-
-    Runs entirely under clone_lock (audit finding C2): without it, this
-    fetch + hard-reset can interleave with an in-flight publish_project's
-    stage/commit/push on the same clone directory, tearing a commit or
-    contending on .git/index.lock.
-    """
-    with clone_lock(url, env):
-        repo = shared_repo_path(url, env)
-        if not (repo / ".git").exists():
-            if ensure_shared_clone(url, env) is not None:
-                return True, ""
-            reason = f"could not clone the repository, check that git can access {url}"
-            logger.warning("refresh_shared_clone: %s", reason)
-            return False, reason
-        if (repo / ".git" / "shallow").exists():
-            run_git(["fetch", "--unshallow", "origin"], cwd=repo, timeout=timeout)
-        ok, out = run_git(["fetch", "origin", "HEAD"], cwd=repo, timeout=timeout)
-        if not ok:
-            reason = out.strip()[:200]
-            logger.warning("refresh_shared_clone: fetch failed for %s: %s", url, reason)
-            return False, reason
-        ok, out = run_git(["reset", "--hard", "FETCH_HEAD"], cwd=repo, timeout=timeout)
-        if not ok:
-            reason = out.strip()[:200]
-            logger.warning("refresh_shared_clone: reset failed for %s: %s", url, reason)
-            return False, reason
-        return True, ""
-
-
-def last_synced_at(url: str, env: dict | None = None) -> float | None:
-    repo = shared_repo_path(url, env)
-    for name in ("FETCH_HEAD", "HEAD"):
-        candidate = repo / ".git" / name
-        try:
-            return candidate.stat().st_mtime
-        except OSError:
-            continue
-    return None
-
-
-MARKER_FILENAME = "quodeq.json"
-FORMAT_NAME = "quodeq-shared-evaluations"
-FORMAT_VERSION = 1
-PUBLISHED_META_FILENAME = "published.json"
-
-_GITIGNORE_CONTENT = "**/evaluation.db\n*.log\n"
-
-
-def check_repo_format(repo_root: Path) -> str:
-    marker = repo_root / MARKER_FILENAME
-    if marker.exists():
-        try:
-            data = json.loads(marker.read_text(encoding="utf-8"))
-        except (OSError, json.JSONDecodeError, UnicodeDecodeError):
-            return "foreign"
-
-        # Marker JSON must be a dict; if not, it's foreign.
-        if not isinstance(data, dict):
-            return "foreign"
-
-        if data.get("format") != FORMAT_NAME:
-            return "foreign"
-
-        # Try to parse version as int; if it fails or is non-numeric, unsupported.
-        try:
-            version = int(data.get("version", 0))
-        except (ValueError, TypeError):
-            return "unsupported_version"
-
-        if version > FORMAT_VERSION:
-            return "unsupported_version"
-        return "ok"
-
-    try:
-        entries = [p for p in repo_root.iterdir() if p.name != ".git"]
-    except OSError:
-        return "foreign"
-    return "empty" if not entries else "foreign"
-
-
-def bootstrap_repo_layout(repo_root: Path) -> None:
-    marker_content = json.dumps({"format": FORMAT_NAME, "version": FORMAT_VERSION}) + "\n"
-    (repo_root / MARKER_FILENAME).write_text(marker_content, encoding="utf-8")
-    (repo_root / ".gitignore").write_text(_GITIGNORE_CONTENT, encoding="utf-8")
-    evaluations = repo_root / "evaluations"
-    evaluations.mkdir(exist_ok=True)
-    (evaluations / ".gitkeep").write_text("", encoding="utf-8")
-
-
-def shared_index_db_path(url: str, env: dict | None = None) -> Path:
-    return shared_cache_dir(url, env) / "index.db"
-
-
-def shared_score_cache_path(url: str, env: dict | None = None) -> Path:
-    return shared_cache_dir(url, env) / "score_cache.db"
-
-
-def sync_shared_index(url: str, env: dict | None = None) -> None:
-    from quodeq.data.sqlite.run_index import open_index, sync_index
-
-    root = shared_evaluations_root(url, env)
-    if not root.is_dir():
-        return
-    db = open_index(shared_index_db_path(url, env))
-    try:
-        sync_index(db, root)
-    finally:
-        db.close()
-
-
-def read_state(url: str, env: dict | None = None) -> str:
-    """State of the local shared clone: ok | empty | foreign |
-    unsupported_version | missing. "empty" (cloned, never published into)
-    is servable -- routes return an empty listing for it."""
-    repo = shared_repo_path(url, env)
-    if not (repo / ".git").exists():
-        return "missing"
-    return check_repo_format(repo)
-
-
-def _read_published_json(entry: Path) -> dict | None:
-    """Read <entry>/published.json, validating both keys.
-
-    Returns None on any failure (missing file, bad JSON, wrong shape) so the
-    caller can fall back to the git-log-derived legacy path -- never raises.
-    """
-    try:
-        data = json.loads((entry / PUBLISHED_META_FILENAME).read_text(encoding="utf-8"))
-    except (OSError, ValueError):
-        return None
-    if not isinstance(data, dict):
-        return None
-    by = data.get("publishedBy")
-    at = data.get("publishedAt")
-    if not isinstance(by, str) or not by:
-        return None
-    if not isinstance(at, int) or isinstance(at, bool):
-        return None
-    return {"publishedBy": by, "publishedAt": at}
-
-
-def published_meta(url: str, env: dict | None = None) -> dict[str, dict]:
-    """Attribution per published project: who published it, and when.
-
-    Prefers the published.json written by stage_project at publish time
-    (see shared_publish.py). Falls back to the legacy git-log-derived
-    lookup for project dirs published before that file existed. The
-    fallback is only correct because the clone is full history (no
-    --depth) -- a shallow clone made `git log -1 -- path` return the tip
-    commit for every path, misattributing every project except the most
-    recently pushed one (audit finding C1).
-    """
-    repo = shared_repo_path(url, env)
-    root = shared_evaluations_root(url, env)
-    result: dict[str, dict] = {}
-    if not root.is_dir():
-        return result
-    for entry in sorted(root.iterdir()):
-        if not entry.is_dir():
-            continue
-        meta = _read_published_json(entry)
-        if meta is not None:
-            result[entry.name] = meta
-            continue
-        ok, out = run_git(
-            ["log", "-1", "--format=%an|%ct", "--", f"evaluations/{entry.name}"],
-            cwd=repo,
-        )
-        if ok and "|" in out:
-            author, _, ts = out.strip().rpartition("|")
-            try:
-                result[entry.name] = {"publishedBy": author, "publishedAt": int(ts)}
-            except ValueError:
-                continue
-    return result
+from quodeq.data.fs.shared_repo import (  # noqa: F401 — re-exported API
+    check_repo_format,
+    clone_lock,
+    ensure_shared_clone,
+    last_synced_at,
+    published_meta,
+    read_state,
+    refresh_shared_clone,
+    remove_clone_dir,
+    shared_cache_dir,
+    shared_evaluations_root,
+    shared_index_db_path,
+    shared_score_cache_path,
+    sync_shared_index,
+    validate_remote_url,
+)
