+++ tests/api/conftest.py
@@ -18,7 +18,7 @@ import pytest
 from quodeq.core.events.models import JudgmentCreatedEvent, JudgmentPayload
 from quodeq.data.events.writer import EventLogWriter
 from quodeq.services.shared_publish import publish_project
-from quodeq.services.shared_repo import (
+from quodeq.data.fs.shared_repo import (
     ensure_shared_clone,
     shared_evaluations_root,
     shared_repo_path,
